chore(filter-field): Refactor unit tests into multiple files.
