+++ libs/barista-components/filter-field/src/filter-field.spec.ts
@@ -19,144 +19,48 @@
 
 import { BACKSPACE, DOWN_ARROW, ENTER, ESCAPE } from '@angular/cdk/keycodes';
 import { OverlayContainer } from '@angular/cdk/overlay';
-import { HttpClientTestingModule } from '@angular/common/http/testing';
-import { Component, DebugElement, NgZone, ViewChild } from '@angular/core';
 import {
   ComponentFixture,
   fakeAsync,
   flush,
   inject,
-  TestBed,
   tick,
 } from '@angular/core/testing';
 import { By } from '@angular/platform-browser';
 import {
   DtFilterField,
   DtFilterFieldChangeEvent,
-  DtFilterFieldDefaultDataSource,
-  DtFilterFieldModule,
-  dtRangeDef,
   DT_FILTER_FIELD_TYPING_DEBOUNCE,
-  getDtFilterFieldRangeNoOperatorsError,
-  DT_FILTER_VALUES_PARSER_CONFIG,
 } from '@dynatrace/barista-components/filter-field';
-import { DtIconModule } from '@dynatrace/barista-components/icon';
 import {
   createComponent,
   dispatchFakeEvent,
   dispatchKeyboardEvent,
   MockNgZone,
-  typeInElement,
-  wrappedErrorMessage,
 } from '@dynatrace/testing/browser';
 import {
   FILTER_FIELD_TEST_DATA_ASYNC,
   FILTER_FIELD_TEST_DATA_SINGLE_DISTINCT,
   FILTER_FIELD_TEST_DATA_SINGLE_OPTION,
 } from '@dynatrace/testing/fixtures';
-import { defaultTagDataForFilterValuesParser } from './filter-field-util';
+import { TEST_DATA_EDITMODE_ASYNC } from './filter-field-edit-mode.spec';
 import {
-  DtFilterValue,
-  DtFilterFieldTagData,
-  isDtAutocompleteValue,
-} from './types';
-
-const TEST_DATA_SUGGESTIONS = {
-  autocomplete: [
-    {
-      name: 'Node',
-      options: [
-        {
-          name: 'Custom Simple Option',
-        },
-        {
-          name: 'Node Label',
-          key: 'MyKey',
-          suggestions: [{ name: 'some cool' }, { name: 'very weird' }],
-          validators: [],
-        },
-      ],
-    },
-  ],
-};
-
-const TEST_DATA_RANGE = {
-  autocomplete: [
-    {
-      name: 'Requests per minute',
-      range: {
-        operators: {
-          range: true,
-          equal: true,
-          greaterThanEqual: true,
-          lessThanEqual: true,
-        },
-        unit: 's',
-      },
-    },
-  ],
-};
-
-const TEST_DATA_EDITMODE = {
-  autocomplete: [
-    {
-      name: 'AUT',
-      autocomplete: [
-        {
-          name: 'Upper Austria',
-          distinct: true,
-          autocomplete: [
-            {
-              name: 'Cities',
-              options: [{ name: 'Linz' }, { name: 'Wels' }, { name: 'Steyr' }],
-            },
-          ],
-        },
-        {
-          name: 'Vienna',
-        },
-      ],
-    },
-    {
-      name: 'USA',
-      autocomplete: [{ name: 'Los Angeles' }, { name: 'San Fran' }],
-    },
-    {
-      name: 'Free',
-      suggestions: [],
-      validators: [],
-    },
-    {
-      name: 'Requests per minute',
-      range: {
-        operators: {
-          range: true,
-          equal: true,
-          greaterThanEqual: true,
-          lessThanEqual: true,
-        },
-        unit: 's',
-      },
-    },
-    {
-      name: 'DE (async)',
-      async: true,
-      autocomplete: [],
-    },
-  ],
-};
-
-export const TEST_DATA_EDITMODE_ASYNC = {
-  name: 'DE (async)',
-  autocomplete: [
-    { name: 'Berlin' },
-    {
-      name: 'München',
-      suggestions: [],
-      validators: [],
-    },
-  ],
-};
+  TEST_DATA_SUGGESTIONS,
+  TEST_DATA_EDITMODE,
+} from './testing/filter-field-test-data';
+import {
+  TestApp,
+  TestAppCustomParserConfig,
+  TestAppCustomParserInput,
+  getOptions,
+  setupFilterFieldTest,
+  getOptionGroups,
+  getClearAll,
+  getFilterTags,
+  getInput,
+  getTagButtons,
+  isClearAllVisible,
+} from './testing/filter-field-test-helpers';
 
 describe('DtFilterField', () => {
   let fixture: ComponentFixture<TestApp>;
@@ -164,74 +68,33 @@ describe('DtFilterField', () => {
   let overlayContainerElement: HTMLElement;
   let filterField: DtFilterField<any>;
   let zone: MockNgZone;
-
-  /**
-   * Simulate zone steps and change detection to advance the
-   * filter field to the next state.
-   */
-  function advanceFilterfieldCycle(
-    simulateMicrotasks: boolean = true,
-    simulateZoneExit: boolean = true,
-  ): void {
-    if (simulateMicrotasks) {
-      zone.simulateMicrotasksEmpty();
-    }
-    if (simulateZoneExit) {
-      zone.simulateZoneExit();
-    }
-    fixture.detectChanges();
-  }
-
-  /**
-   * Gets the nth option from the passed overlay container, clicks it and advances
-   * the filter field cycle.
-   */
-  function getAndClickOption(
+  let advanceFilterfieldCycle: (
+    simulateMicrotasks?: boolean,
+    simulateZoneExit?: boolean,
+  ) => void;
+  let getAndClickOption: (
     optionOverlayContainer: HTMLElement,
     nthOption: number,
-  ): void {
-    let options = getOptions(optionOverlayContainer);
-    const selectedOption = options[nthOption];
-    selectedOption.click();
-    advanceFilterfieldCycle();
-  }
+  ) => void;
+  let typeIntoFilterElement: (inputString: string) => void;
 
-  /**
-   * Types the passed value into the filter field input element,
-   * waits for the debounce time.
-   */
-  function typeIntoFilterElement(inputString: string): void {
-    const inputEl = fixture.debugElement.query(By.css('input')).nativeElement;
-    typeInElement(inputString, inputEl);
-    tick(DT_FILTER_FIELD_TYPING_DEBOUNCE);
-  }
-
-  beforeEach(fakeAsync(() => {
-    TestBed.configureTestingModule({
-      imports: [
-        HttpClientTestingModule,
-        DtIconModule.forRoot({ svgIconLocation: `{{name}}.svg` }),
-        DtFilterFieldModule,
-      ],
-      declarations: [
-        TestApp,
-        TestAppCustomParserConfig,
-        TestAppCustomParserInput,
-      ],
-      providers: [
-        { provide: NgZone, useFactory: () => (zone = new MockNgZone()) },
-      ],
-    }).compileComponents();
+  beforeEach(() => {
+    ({
+      fixture,
+      filterField,
+      zone,
+      overlayContainer,
+      overlayContainerElement,
+      advanceFilterfieldCycle,
+      getAndClickOption,
+      typeIntoFilterElement,
+    } = setupFilterFieldTest());
 
     inject([OverlayContainer], (oc: OverlayContainer) => {
       overlayContainer = oc;
       overlayContainerElement = oc.getContainerElement();
     })();
-
-    fixture = createComponent(TestApp);
-    filterField = fixture.debugElement.query(By.directive(DtFilterField))
-      .componentInstance;
-  }));
+  });
 
   afterEach(inject(
     [OverlayContainer],
@@ -1005,731 +868,6 @@ describe('DtFilterField', () => {
     });
   });
 
-  describe('with range option', () => {
-    beforeEach(() => {
-      fixture.componentInstance.dataSource.data = TEST_DATA_RANGE;
-      fixture.detectChanges();
-
-      // Focus the filter field.
-      filterField.focus();
-      advanceFilterfieldCycle();
-    });
-
-    it('should open the range overlay if the range option is selected', () => {
-      let filterFieldRangeElements = getFilterFieldRange(
-        overlayContainerElement,
-      );
-      expect(filterFieldRangeElements.length).toBe(0);
-
-      getAndClickOption(overlayContainerElement, 0);
-
-      filterFieldRangeElements = getFilterFieldRange(overlayContainerElement);
-
-      expect(filterFieldRangeElements.length).toBe(1);
-    });
-
-    it('should have only range and greater-equal operators enabled', () => {
-      // Change the datasource at runtime
-      fixture.componentInstance.dataSource.data = {
-        autocomplete: [
-          {
-            name: 'Requests per minute',
-            range: {
-              operators: {
-                range: true,
-                greaterThanEqual: true,
-              },
-              unit: 's',
-            },
-          },
-        ],
-      };
-      fixture.detectChanges();
-
-      // open the range overlay
-      filterField.focus();
-      advanceFilterfieldCycle();
-      const options = getOptions(overlayContainerElement);
-      options[0].click();
-
-      advanceFilterfieldCycle();
-
-      const operatorButtonElements = getOperatorButtonGroupItems(
-        overlayContainerElement,
-      );
-      expect(operatorButtonElements.length).toBe(2);
-      expect(operatorButtonElements[0].textContent).toBe('Range');
-      expect(operatorButtonElements[1].textContent).toBe('≥');
-    });
-
-    it('should have only one operator enabled', () => {
-      // Change the datasource at runtime
-      fixture.componentInstance.dataSource.data = {
-        autocomplete: [
-          {
-            name: 'Requests per minute',
-            range: {
-              operators: {
-                greaterThanEqual: true,
-              },
-              unit: 's',
-            },
-          },
-        ],
-      };
-      fixture.detectChanges();
-
-      // open the range overlay
-      filterField.focus();
-      advanceFilterfieldCycle();
-
-      getAndClickOption(overlayContainerElement, 0);
-
-      const operatorButtonElements = getOperatorButtonGroupItems(
-        overlayContainerElement,
-      );
-      expect(operatorButtonElements.length).toBe(1);
-      expect(operatorButtonElements[0].textContent).toBe('≥');
-    });
-
-    describe('opened', () => {
-      beforeEach(() => {
-        // Open the filter-field-range overlay.
-        getAndClickOption(overlayContainerElement, 0);
-      });
-
-      it('should set the focus onto the first button-group-item by default', () => {
-        const firstButtonElement = overlayContainerElement.querySelector(
-          'dt-button-group-item',
-        );
-        expect(document.activeElement).toBe(firstButtonElement);
-      });
-
-      it('should have all operators enabled', () => {
-        const operatorButtonElements = getOperatorButtonGroupItems(
-          overlayContainerElement,
-        );
-        expect(operatorButtonElements.length).toBe(4);
-      });
-
-      it('should throw an error if there is no operator defined', fakeAsync(() => {
-        // Change the datasource at runtime
-        expect(() => {
-          fixture.componentInstance.dataSource.data = {
-            autocomplete: [
-              {
-                name: 'Requests per minute',
-                range: {
-                  operators: {},
-                  unit: 's',
-                },
-              },
-            ],
-          };
-          flush();
-        }).toThrowError(
-          wrappedErrorMessage(getDtFilterFieldRangeNoOperatorsError()),
-        );
-      }));
-
-      it('should throw', () => {
-        expect(() => {
-          dtRangeDef({}, null, false, false, false, false, 's', false);
-        }).toThrowError(
-          wrappedErrorMessage(getDtFilterFieldRangeNoOperatorsError()),
-        );
-      });
-
-      it('should show two input fields when operator range is selected', () => {
-        const inputFieldsElements = getRangeInputFields(
-          overlayContainerElement,
-        );
-        expect(inputFieldsElements.length).toBe(2);
-      });
-
-      it('should show only one input field when the operator is not range', () => {
-        const operatorButtonElements = getOperatorButtonGroupItems(
-          overlayContainerElement,
-        );
-        operatorButtonElements[2].click();
-        fixture.detectChanges();
-
-        const inputFieldsElements = getRangeInputFields(
-          overlayContainerElement,
-        );
-        expect(inputFieldsElements.length).toBe(1);
-      });
-
-      it('should keep the apply button disabled until both entries are valid', () => {
-        const inputFieldsElements = getRangeInputFields(
-          overlayContainerElement,
-        );
-
-        let rangeApplyButton = getRangeApplyButton(overlayContainerElement)[0];
-        expect(rangeApplyButton.getAttribute('disabled')).toBeDefined();
-
-        typeInElement('15', inputFieldsElements[0]);
-        fixture.detectChanges();
-
-        rangeApplyButton = getRangeApplyButton(overlayContainerElement)[0];
-        expect(rangeApplyButton.getAttribute('disabled')).toBeDefined();
-
-        typeInElement('25', inputFieldsElements[1]);
-        fixture.detectChanges();
-
-        rangeApplyButton = getRangeApplyButton(overlayContainerElement)[0];
-        expect(rangeApplyButton.getAttribute('disabled')).toBeNull();
-      });
-
-      it('should keep the apply button disabled until the one required is valid', () => {
-        const operatorButtonElements = getOperatorButtonGroupItems(
-          overlayContainerElement,
-        );
-
-        operatorButtonElements[2].click();
-        fixture.detectChanges();
-
-        const inputFieldsElements = getRangeInputFields(
-          overlayContainerElement,
-        );
-
-        let rangeApplyButton = getRangeApplyButton(overlayContainerElement)[0];
-        expect(rangeApplyButton.getAttribute('disabled')).toBeDefined();
-
-        typeInElement('15', inputFieldsElements[0]);
-        fixture.detectChanges();
-
-        rangeApplyButton = getRangeApplyButton(overlayContainerElement)[0];
-        expect(rangeApplyButton.getAttribute('disabled')).toBeNull();
-      });
-
-      it('should close the range after the range-filter is submitted', () => {
-        const inputFieldsElements = getRangeInputFields(
-          overlayContainerElement,
-        );
-
-        typeInElement('15', inputFieldsElements[0]);
-        typeInElement('25', inputFieldsElements[1]);
-        fixture.detectChanges();
-
-        const rangeApplyButton = getRangeApplyButton(
-          overlayContainerElement,
-        )[0];
-        rangeApplyButton.click();
-
-        const rangeOverlay = getFilterFieldRange(overlayContainerElement);
-        expect(rangeOverlay.length).toBe(0);
-      });
-
-      it('should have the tag-filter committed in the filterfield and formatted for range', () => {
-        const inputFieldsElements = getRangeInputFields(
-          overlayContainerElement,
-        );
-
-        typeInElement('15', inputFieldsElements[0]);
-        typeInElement('25', inputFieldsElements[1]);
-        fixture.detectChanges();
-
-        const rangeApplyButton = getRangeApplyButton(
-          overlayContainerElement,
-        )[0];
-        rangeApplyButton.click();
-        fixture.detectChanges();
-
-        const tags = getFilterTags(fixture);
-        expect(tags[0].key).toBe('Requests per minute');
-        expect(tags[0].separator).toBe(':');
-        expect(tags[0].value.trim()).toBe('15s - 25s');
-      });
-
-      it('should have the tag-filter committed in the filterfield and formatted for greater-than operator', () => {
-        const operatorButtonElements = getOperatorButtonGroupItems(
-          overlayContainerElement,
-        );
-
-        operatorButtonElements[2].click();
-        fixture.detectChanges();
-
-        const inputFieldsElements = getRangeInputFields(
-          overlayContainerElement,
-        );
-
-        typeInElement('15', inputFieldsElements[0]);
-        fixture.detectChanges();
-
-        const rangeApplyButton = getRangeApplyButton(
-          overlayContainerElement,
-        )[0];
-        rangeApplyButton.click();
-        fixture.detectChanges();
-
-        const tags = getFilterTags(fixture);
-        expect(tags[0].key).toBe('Requests per minute');
-        expect(tags[0].separator).toBe('≥');
-        expect(tags[0].value.trim()).toBe('15s');
-      });
-
-      it('should close the filter-range when using the keyboard ESC', () => {
-        const operatorButtonElements = getOperatorButtonGroupItems(
-          overlayContainerElement,
-        );
-        dispatchKeyboardEvent(operatorButtonElements[0], 'keydown', ESCAPE);
-        fixture.detectChanges();
-
-        const rangeOverlay = getFilterFieldRange(overlayContainerElement);
-        expect(rangeOverlay.length).toBe(0);
-      });
-
-      it('should reapply previously set range operator and values when editing a range filter', () => {
-        const inputFieldsElements = getRangeInputFields(
-          overlayContainerElement,
-        );
-
-        typeInElement('15', inputFieldsElements[0]);
-        typeInElement('25', inputFieldsElements[1]);
-        fixture.detectChanges();
-
-        const rangeApplyButton = getRangeApplyButton(
-          overlayContainerElement,
-        )[0];
-        rangeApplyButton.click();
-        fixture.detectChanges();
-
-        const tagLabel = fixture.debugElement.queryAll(
-          By.css('.dt-filter-field-tag-label'),
-        )[0];
-        tagLabel.nativeElement.click();
-
-        advanceFilterfieldCycle();
-
-        // expect the range to open again
-        const filterFieldRangeElements = getFilterFieldRange(
-          overlayContainerElement,
-        );
-        expect(filterFieldRangeElements.length).toBe(1);
-
-        // expect the values to be filled
-        expect(filterField._filterfieldRange._selectedOperator).toEqual(
-          'range',
-        );
-        expect(filterField._filterfieldRange._valueFrom).toEqual('15');
-        expect(filterField._filterfieldRange._valueTo).toEqual('25');
-      });
-
-      it('should reapply previously set greater-equal operator and value when editing a range filter', () => {
-        const operatorButtonElements = getOperatorButtonGroupItems(
-          overlayContainerElement,
-        );
-        operatorButtonElements[2].click();
-        fixture.detectChanges();
-
-        const inputFieldsElements = getRangeInputFields(
-          overlayContainerElement,
-        );
-        typeInElement('27', inputFieldsElements[0]);
-        fixture.detectChanges();
-
-        const rangeApplyButton = getRangeApplyButton(
-          overlayContainerElement,
-        )[0];
-        rangeApplyButton.click();
-        fixture.detectChanges();
-
-        const tagLabel = fixture.debugElement.queryAll(
-          By.css('.dt-filter-field-tag-label'),
-        )[0];
-        tagLabel.nativeElement.click();
-
-        advanceFilterfieldCycle();
-
-        // expect the range to open again
-        const filterFieldRangeElements = getFilterFieldRange(
-          overlayContainerElement,
-        );
-        expect(filterFieldRangeElements.length).toBe(1);
-
-        // expect the values to be filled
-        expect(filterField._filterfieldRange._selectedOperator).toEqual(
-          'greater-equal',
-        );
-        expect(filterField._filterfieldRange._valueFrom).toEqual('27');
-      });
-    });
-  });
-
-  describe('edit mode', () => {
-    beforeEach(() => {
-      fixture.componentInstance.dataSource.data = TEST_DATA_EDITMODE;
-      advanceFilterfieldCycle();
-
-      // Autocomplete filter for AUT -> Upper Austria -> Cities -> Linz
-      const autocompleteFilter = [
-        TEST_DATA_EDITMODE.autocomplete[0],
-        (TEST_DATA_EDITMODE as any).autocomplete[0].autocomplete[0],
-        (TEST_DATA_EDITMODE as any).autocomplete[0].autocomplete[0]
-          .autocomplete[0].options[0],
-      ];
-      // Custom free text for Free -> Custom free text
-      const freeTextFilter = [
-        TEST_DATA_EDITMODE.autocomplete[2],
-        'Custom free text',
-      ];
-
-      const rangeFilter = [
-        TEST_DATA_EDITMODE.autocomplete[3],
-        { operator: 'range', unit: 's', range: [15, 80] },
-      ];
-
-      // Set filters as a starting point
-      filterField.filters = [autocompleteFilter, freeTextFilter, rangeFilter];
-      fixture.detectChanges();
-    });
-
-    it('should have the correct filters set as a starting point', () => {
-      const tags = getFilterTags(fixture);
-
-      expect(tags[0].key).toBe('AUT');
-      expect(tags[0].separator).toBe(':');
-      expect(tags[0].value).toBe('Linz');
-
-      expect(tags[1].key).toBe('Free');
-      expect(tags[1].separator).toBe('~');
-      expect(tags[1].value).toBe('Custom free text');
-
-      expect(tags[2].key).toBe('Requests per minute');
-      expect(tags[2].separator).toBe(':');
-      expect(tags[2].value).toBe('15s - 80s');
-    });
-
-    it('should reset the autocomplete filter when not changing anything and cancelling by mouse', () => {
-      const tags = fixture.debugElement.queryAll(
-        By.css('.dt-filter-field-tag-label'),
-      );
-      tags[0].nativeElement.click();
-      advanceFilterfieldCycle();
-
-      let options = getOptions(overlayContainerElement);
-      // Make sure that the autocomplete actually opened.
-      expect(options.length).toBe(2);
-
-      dispatchFakeEvent(document, 'click');
-      advanceFilterfieldCycle();
-
-      // Make sure the autocomplete closed again.
-      options = getOptions(overlayContainerElement);
-      expect(options.length).toBe(0);
-
-      // Read the filters again and make expectations
-      const filterTags = getFilterTags(fixture);
-
-      expect(filterTags[0].key).toBe('AUT');
-      expect(filterTags[0].separator).toBe(':');
-      expect(filterTags[0].value).toBe('Linz');
-
-      expect(filterTags[1].key).toBe('Free');
-      expect(filterTags[1].separator).toBe('~');
-      expect(filterTags[1].value).toBe('Custom free text');
-
-      expect(filterTags[2].key).toBe('Requests per minute');
-      expect(filterTags[2].separator).toBe(':');
-      expect(filterTags[2].value).toBe('15s - 80s');
-    });
-
-    it('should reset the autocomplete filter when not changing anything and cancelling by keyboard', () => {
-      const tags = fixture.debugElement.queryAll(
-        By.css('.dt-filter-field-tag-label'),
-      );
-      tags[0].nativeElement.click();
-      advanceFilterfieldCycle();
-
-      let options = getOptions(overlayContainerElement);
-      // Make sure that the autocomplete actually opened.
-      expect(options.length).toBe(2);
-
-      // Cancel editmode with keyboard
-      const inputfield = getInput(fixture);
-      dispatchKeyboardEvent(inputfield, 'keydown', ESCAPE);
-      advanceFilterfieldCycle();
-
-      // Make sure the autocomplete closed again.
-      options = getOptions(overlayContainerElement);
-      expect(options.length).toBe(0);
-
-      // Read the filters again and make expectations
-      const filterTags = getFilterTags(fixture);
-
-      expect(filterTags[0].key).toBe('AUT');
-      expect(filterTags[0].separator).toBe(':');
-      expect(filterTags[0].value).toBe('Linz');
-
-      expect(filterTags[1].key).toBe('Free');
-      expect(filterTags[1].separator).toBe('~');
-      expect(filterTags[1].value).toBe('Custom free text');
-
-      expect(filterTags[2].key).toBe('Requests per minute');
-      expect(filterTags[2].separator).toBe(':');
-      expect(filterTags[2].value).toBe('15s - 80s');
-    });
-
-    it('should reset the freetext filter when not changing anything and cancelling by mouse', () => {
-      const tags = fixture.debugElement.queryAll(
-        By.css('.dt-filter-field-tag-label'),
-      );
-      tags[1].nativeElement.click();
-      advanceFilterfieldCycle();
-
-      dispatchFakeEvent(document, 'click');
-      advanceFilterfieldCycle();
-
-      // Read the filters again and make expectations
-      const filterTags = getFilterTags(fixture);
-
-      expect(filterTags[0].key).toBe('AUT');
-      expect(filterTags[0].separator).toBe(':');
-      expect(filterTags[0].value).toBe('Linz');
-
-      expect(filterTags[1].key).toBe('Free');
-      expect(filterTags[1].separator).toBe('~');
-      expect(filterTags[1].value).toBe('Custom free text');
-
-      expect(filterTags[2].key).toBe('Requests per minute');
-      expect(filterTags[2].separator).toBe(':');
-      expect(filterTags[2].value).toBe('15s - 80s');
-    });
-
-    it('should reset the input value when editing the freetext, typing something (but not commiting the filter) and then cancelling', () => {
-      const tags = fixture.debugElement.queryAll(
-        By.css('.dt-filter-field-tag-label'),
-      );
-      tags[1].nativeElement.click();
-      advanceFilterfieldCycle();
-
-      const inputfield = getInput(fixture);
-      typeInElement('something else', inputfield);
-      fixture.detectChanges();
-
-      dispatchFakeEvent(document, 'click');
-      advanceFilterfieldCycle();
-
-      // Read the filters again and make expectations
-      const filterTags = getFilterTags(fixture);
-
-      expect(filterTags[0].key).toBe('AUT');
-      expect(filterTags[0].separator).toBe(':');
-      expect(filterTags[0].value).toBe('Linz');
-
-      expect(filterTags[1].key).toBe('Free');
-      expect(filterTags[1].separator).toBe('~');
-      expect(filterTags[1].value).toBe('Custom free text');
-
-      expect(filterTags[2].key).toBe('Requests per minute');
-      expect(filterTags[2].separator).toBe(':');
-      expect(filterTags[2].value).toBe('15s - 80s');
-
-      expect(inputfield.value).toBe('');
-    });
-
-    it('should reset the freetext filter when not changing anything and cancelling by keyboard', () => {
-      const tags = fixture.debugElement.queryAll(
-        By.css('.dt-filter-field-tag-label'),
-      );
-      tags[1].nativeElement.click();
-      advanceFilterfieldCycle();
-
-      // Cancel editmode with keyboard
-      const inputfield = getInput(fixture);
-      dispatchKeyboardEvent(inputfield, 'keydown', ESCAPE);
-      advanceFilterfieldCycle();
-
-      // Read the filters again and make expectations
-      const filterTags = getFilterTags(fixture);
-
-      expect(filterTags[0].key).toBe('AUT');
-      expect(filterTags[0].separator).toBe(':');
-      expect(filterTags[0].value).toBe('Linz');
-
-      expect(filterTags[1].key).toBe('Free');
-      expect(filterTags[1].separator).toBe('~');
-      expect(filterTags[1].value).toBe('Custom free text');
-
-      expect(filterTags[2].key).toBe('Requests per minute');
-      expect(filterTags[2].separator).toBe(':');
-      expect(filterTags[2].value).toBe('15s - 80s');
-    });
-
-    it('should reset the range filter when not changing anything and cancelling by mouse', () => {
-      const tags = fixture.debugElement.queryAll(
-        By.css('.dt-filter-field-tag-label'),
-      );
-      tags[2].nativeElement.click();
-      advanceFilterfieldCycle();
-
-      // Expect the range filter to be open
-      let filterfield = getFilterFieldRange(overlayContainerElement);
-      expect(filterfield.length).toBe(1);
-
-      dispatchFakeEvent(document, 'click');
-      advanceFilterfieldCycle();
-
-      // Expect the range filter to be closed again
-      filterfield = getFilterFieldRange(overlayContainerElement);
-      expect(filterfield.length).toBe(0);
-
-      // Read the filters again and make expectations
-      const filterTags = getFilterTags(fixture);
-
-      expect(filterTags[0].key).toBe('AUT');
-      expect(filterTags[0].separator).toBe(':');
-      expect(filterTags[0].value).toBe('Linz');
-
-      expect(filterTags[1].key).toBe('Free');
-      expect(filterTags[1].separator).toBe('~');
-      expect(filterTags[1].value).toBe('Custom free text');
-
-      expect(filterTags[2].key).toBe('Requests per minute');
-      expect(filterTags[2].separator).toBe(':');
-      expect(filterTags[2].value).toBe('15s - 80s');
-    });
-
-    it('should reset the range filter when not changing anything and cancelling by keyboard', () => {
-      const tags = fixture.debugElement.queryAll(
-        By.css('.dt-filter-field-tag-label'),
-      );
-      tags[2].nativeElement.click();
-      advanceFilterfieldCycle();
-
-      // Expect the range filter to be open
-      let filterfield = getFilterFieldRange(overlayContainerElement);
-      expect(filterfield.length).toBe(1);
-
-      // Cancel editmode with keyboard
-      const inputfields = getRangeInputFields(overlayContainerElement);
-      dispatchKeyboardEvent(inputfields[0], 'keydown', ESCAPE);
-
-      advanceFilterfieldCycle();
-
-      // Expect the range filter to be closed again
-      filterfield = getFilterFieldRange(overlayContainerElement);
-      expect(filterfield.length).toBe(0);
-
-      // Read the filters again and make expectations
-      const filterTags = getFilterTags(fixture);
-
-      expect(filterTags[0].key).toBe('AUT');
-      expect(filterTags[0].separator).toBe(':');
-      expect(filterTags[0].value).toBe('Linz');
-
-      expect(filterTags[1].key).toBe('Free');
-      expect(filterTags[1].separator).toBe('~');
-      expect(filterTags[1].value).toBe('Custom free text');
-
-      expect(filterTags[2].key).toBe('Requests per minute');
-      expect(filterTags[2].separator).toBe(':');
-      expect(filterTags[2].value).toBe('15s - 80s');
-    });
-
-    it('should not reset the range filter when focusing the range input element', () => {
-      const tags = fixture.debugElement.queryAll(
-        By.css('.dt-filter-field-tag-label'),
-      );
-      tags[2].nativeElement.click();
-      advanceFilterfieldCycle();
-
-      const inputfields = getRangeInputFields(overlayContainerElement);
-      inputfields[0].click();
-      advanceFilterfieldCycle();
-
-      expect(filterField.filters).toHaveLength(3);
-      // Range filter should have set only the root filter (range def)
-      expect(filterField.filters[2]).toHaveLength(1);
-    });
-
-    it('should make the edit to the first tag', () => {
-      const tags = fixture.debugElement.queryAll(
-        By.css('.dt-filter-field-tag-label'),
-      );
-      tags[0].nativeElement.click();
-      advanceFilterfieldCycle();
-
-      const options = getOptions(overlayContainerElement);
-      // Make sure that the autocomplete actually opened.
-      expect(options.length).toBe(2);
-
-      // Select Vienna
-      options[1].click();
-      advanceFilterfieldCycle();
-
-      // Read the filters again and make expectations
-      const filterTags = getFilterTags(fixture);
-
-      expect(filterTags[0].key).toBe('AUT');
-      expect(filterTags[0].separator).toBe(':');
-      expect(filterTags[0].value).toBe('Vienna');
-
-      expect(filterTags[1].key).toBe('Free');
-      expect(filterTags[1].separator).toBe('~');
-      expect(filterTags[1].value).toBe('Custom free text');
-
-      expect(filterTags[2].key).toBe('Requests per minute');
-      expect(filterTags[2].separator).toBe(':');
-      expect(filterTags[2].value).toBe('15s - 80s');
-    });
-
-    it('should not reset anything when a filter is deleted in edit mode and an outside click is triggered', () => {
-      let tags = fixture.debugElement.queryAll(
-        By.css('.dt-filter-field-tag-label'),
-      );
-
-      expect(tags.length).toBe(3);
-
-      tags[0].nativeElement.click();
-      advanceFilterfieldCycle();
-
-      const inputEl = getInput(fixture);
-      dispatchKeyboardEvent(inputEl, 'keydown', BACKSPACE);
-      fixture.detectChanges();
-      tags = fixture.debugElement.queryAll(
-        By.css('.dt-filter-field-tag-label'),
-      );
-
-      expect(tags.length).toBe(2);
-
-      dispatchFakeEvent(document, 'click');
-      fixture.detectChanges();
-      tags = fixture.debugElement.queryAll(
-        By.css('.dt-filter-field-tag-label'),
-      );
-
-      expect(tags.length).toBe(2);
-    });
-
-    it('should emit a filterchange event when the edit of a range is completed', () => {
-      let filterChangeEvent: DtFilterFieldChangeEvent<any> | undefined;
-
-      fixture.componentInstance.dataSource.data = FILTER_FIELD_TEST_DATA_SINGLE_OPTION;
-      const sub = filterField.filterChanges.subscribe(
-        (ev) => (filterChangeEvent = ev),
-      );
-
-      const tags = fixture.debugElement.queryAll(
-        By.css('.dt-filter-field-tag-label'),
-      );
-
-      tags[2].nativeElement.click();
-      advanceFilterfieldCycle();
-
-      const applyButton = getRangeApplyButton(overlayContainerElement)[0];
-      applyButton.click();
-      fixture.detectChanges();
-
-      expect(filterChangeEvent).toBeDefined();
-      expect(filterChangeEvent!.added.length).toBe(1);
-      expect(filterChangeEvent!.removed.length).toBe(0);
-
-      sub.unsubscribe();
-    });
-  });
-
   describe('programmatic setting', () => {
     beforeEach(() => {
       fixture.componentInstance.dataSource.data = TEST_DATA_EDITMODE;
@@ -2412,210 +1550,3 @@ describe('DtFilterField', () => {
     });
   });
 });
-
-function getOptions(overlayContainerElement: HTMLElement): HTMLElement[] {
-  return Array.from(overlayContainerElement.querySelectorAll('.dt-option'));
-}
-
-function getOptionGroups(overlayContainerElement: HTMLElement): HTMLElement[] {
-  return Array.from(overlayContainerElement.querySelectorAll('.dt-optgroup'));
-}
-
-function getFilterFieldRange(
-  overlayContainerElement: HTMLElement,
-): HTMLElement[] {
-  return Array.from(
-    overlayContainerElement.querySelectorAll('.dt-filter-field-range-panel'),
-  );
-}
-
-function getOperatorButtonGroupItems(
-  overlayContainerElement: HTMLElement,
-): HTMLElement[] {
-  return Array.from(
-    overlayContainerElement.querySelectorAll('dt-button-group-item'),
-  );
-}
-
-function getRangeInputFields(
-  overlayContainerElement: HTMLElement,
-): HTMLInputElement[] {
-  return Array.from(
-    overlayContainerElement.querySelectorAll('.dt-filter-field-range-input'),
-  );
-}
-
-function getRangeApplyButton(
-  overlayContainerElement: HTMLElement,
-): HTMLElement[] {
-  return Array.from(
-    overlayContainerElement.querySelectorAll('.dt-filter-field-range-apply'),
-  );
-}
-
-interface FilterTagTestData {
-  ele: DebugElement;
-  key: string;
-  separator: string;
-  value: string;
-  removeButton: HTMLElement;
-}
-
-// tslint:disable-next-line:no-any
-function getFilterTags(
-  fixture: ComponentFixture<any>,
-): Array<FilterTagTestData> {
-  return Array.from(
-    fixture.debugElement.queryAll(By.css('.dt-filter-field-tag')),
-  ).map((ele) => {
-    const key: HTMLElement = ele.nativeElement.querySelector(
-      '.dt-filter-field-tag-key',
-    );
-    const separator =
-      key && key.getAttribute('data-separator')
-        ? key.getAttribute('data-separator')!
-        : '';
-
-    // Get the value of the filter element, if no filter value element is rendered
-    // assume an empty string.
-    const valueElement = ele.nativeElement.querySelector(
-      '.dt-filter-field-tag-value',
-    );
-    const value = valueElement ? valueElement.textContent : '';
-
-    return {
-      ele,
-      key: key && key.textContent ? key.textContent : '',
-      separator,
-      value,
-      removeButton: ele.nativeElement.querySelector(
-        '.dt-filter-field-tag-button',
-      ),
-    };
-  });
-}
-
-function getTagButtons(
-  tag: FilterTagTestData,
-): { label: HTMLButtonElement; deleteButton: HTMLButtonElement } {
-  const tagNative = tag.ele.nativeElement;
-  const label = tagNative.querySelector('.dt-filter-field-tag-label');
-  const deleteButton = tagNative.querySelector('.dt-filter-field-tag-button');
-  return { label, deleteButton };
-}
-
-function getInput(fixture: ComponentFixture<any>): HTMLInputElement {
-  return fixture.debugElement.query(By.css('.dt-filter-field-input'))
-    .nativeElement;
-}
-
-/** Get the clearAll button. */
-function getClearAll(fixture: ComponentFixture<any>): HTMLButtonElement | null {
-  const dbgEl = fixture.debugElement.query(
-    By.css('.dt-filter-field-clear-all-button'),
-  );
-  return dbgEl ? dbgEl.nativeElement : null;
-}
-
-/** Get the clearAll button and evaluate if it is visible or not. */
-function isClearAllVisible(fixture: ComponentFixture<any>): boolean {
-  const clearAll = getClearAll(fixture);
-  return (
-    clearAll !== null &&
-    !clearAll.classList.contains('dt-filter-field-clear-all-button-hidden')
-  );
-}
-
-export function customParser(
-  filterValues: DtFilterValue[],
-  editable?: boolean,
-  deletable?: boolean,
-): DtFilterFieldTagData | null {
-  const tagData = defaultTagDataForFilterValuesParser(
-    filterValues,
-    editable,
-    deletable,
-  );
-  if (tagData) {
-    let isFirstValue = true;
-    for (const filterValue of filterValues) {
-      if (isDtAutocompleteValue(filterValue)) {
-        if (isFirstValue && filterValues.length > 1) {
-          tagData.key = filterValue.option.viewValue;
-        } else if (filterValue.autocomplete && filterValues.length > 1) {
-          tagData.key = tagData.key + '.' + filterValue.option.viewValue;
-        }
-        isFirstValue = false;
-      }
-    }
-  }
-  return tagData;
-}
-
-@Component({
-  selector: 'test-app',
-  template: `
-    <dt-filter-field
-      [dataSource]="dataSource"
-      [label]="label"
-      [clearAllLabel]="clearAllLabel"
-    ></dt-filter-field>
-  `,
-})
-export class TestApp {
-  // tslint:disable-next-line:no-any
-  dataSource = new DtFilterFieldDefaultDataSource(FILTER_FIELD_TEST_DATA_ASYNC);
-
-  label = 'Filter by';
-  clearAllLabel = 'Clear all';
-
-  @ViewChild(DtFilterField) filterField: DtFilterField<any>;
-}
-
-@Component({
-  selector: 'test-app',
-  template: `
-    <dt-filter-field
-      [dataSource]="dataSource"
-      [label]="label"
-      [clearAllLabel]="clearAllLabel"
-    ></dt-filter-field>
-  `,
-  providers: [
-    {
-      provide: DT_FILTER_VALUES_PARSER_CONFIG,
-      useValue: customParser,
-    },
-  ],
-})
-export class TestAppCustomParserConfig {
-  // tslint:disable-next-line:no-any
-  dataSource = new DtFilterFieldDefaultDataSource(FILTER_FIELD_TEST_DATA_ASYNC);
-
-  label = 'Filter by';
-  clearAllLabel = 'Clear all';
-
-  @ViewChild(DtFilterField) filterField: DtFilterField<any>;
-}
-
-@Component({
-  selector: 'test-app',
-  template: `
-    <dt-filter-field
-      [dataSource]="dataSource"
-      [label]="label"
-      [clearAllLabel]="clearAllLabel"
-      [customTagParser]="parserFn"
-    ></dt-filter-field>
-  `,
-})
-export class TestAppCustomParserInput {
-  // tslint:disable-next-line:no-any
-  dataSource = new DtFilterFieldDefaultDataSource(FILTER_FIELD_TEST_DATA_ASYNC);
-
-  label = 'Filter by';
-  clearAllLabel = 'Clear all';
-  parserFn = customParser;
-
-  @ViewChild(DtFilterField) filterField: DtFilterField<any>;
-}
