+++ libs/barista-components/filter-field/src/testing/filter-field-test-data.ts
@@ -1,0 +1,100 @@
+/**
+ * @license
+ * Copyright 2020 Dynatrace LLC
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+export const TEST_DATA_SUGGESTIONS = {
+  autocomplete: [
+    {
+      name: 'Node',
+      options: [
+        {
+          name: 'Custom Simple Option',
+        },
+        {
+          name: 'Node Label',
+          key: 'MyKey',
+          suggestions: [{ name: 'some cool' }, { name: 'very weird' }],
+          validators: [],
+        },
+      ],
+    },
+  ],
+};
+
+export const TEST_DATA_RANGE = {
+  autocomplete: [
+    {
+      name: 'Requests per minute',
+      range: {
+        operators: {
+          range: true,
+          equal: true,
+          greaterThanEqual: true,
+          lessThanEqual: true,
+        },
+        unit: 's',
+      },
+    },
+  ],
+};
+
+export const TEST_DATA_EDITMODE = {
+  autocomplete: [
+    {
+      name: 'AUT',
+      autocomplete: [
+        {
+          name: 'Upper Austria',
+          distinct: true,
+          autocomplete: [
+            {
+              name: 'Cities',
+              options: [{ name: 'Linz' }, { name: 'Wels' }, { name: 'Steyr' }],
+            },
+          ],
+        },
+        {
+          name: 'Vienna',
+        },
+      ],
+    },
+    {
+      name: 'USA',
+      autocomplete: [{ name: 'Los Angeles' }, { name: 'San Fran' }],
+    },
+    {
+      name: 'Free',
+      suggestions: [],
+      validators: [],
+    },
+    {
+      name: 'Requests per minute',
+      range: {
+        operators: {
+          range: true,
+          equal: true,
+          greaterThanEqual: true,
+          lessThanEqual: true,
+        },
+        unit: 's',
+      },
+    },
+    {
+      name: 'DE (async)',
+      async: true,
+      autocomplete: [],
+    },
+  ],
+};
